+++ cmd/api/user/main.go
@@ -4,7 +4,7 @@ import (
 	"flag"
 	"fmt"
 	"github.com/gin-gonic/gin"
-	"gophr.v2/config"
+	"gophr.v2/config/configutil"
 	"gophr.v2/user/api/v1/http"
 	"gophr.v2/user/repository"
 	"gophr.v2/user/service"
@@ -14,7 +14,7 @@ var port = flag.String("port", "4401", "Port")
 
 func main() {
 	flag.Parse()
-	conf := config.Initialize()
+	conf := configutil.Initialize()
 
 	repo, closer := repository.Get(conf, repository.MySQLRepo)
 	defer noOpCloser(closer)
