+++ config/configutil/util_integ_test.go
@@ -1,11 +1,9 @@
 // +Build integration
 
-package config_test
+package configutil_test
 
 import (
-	"github.com/magiconair/properties/assert"
 	"github.com/stretchr/testify/require"
-	"gophr.v2/config"
 	"io/ioutil"
 	"os"
 	"path/filepath"
@@ -35,10 +33,3 @@ mysql:
 	}
 }
 
-func TestConfig(t *testing.T) {
-	closer := setup(t)
-	defer closer()
-	config, err := config.LoadDefault(config.DevelopmentEnv)
-	require.NoError(t, err)
-	assert.Equal(t, "127.0.0.1", config.MySQL.Host)
-}
